+++ app/tests/pages_tests/test_substitutions.py
@@ -1,7 +1,9 @@
 import pytest
 from django.utils.safestring import SafeString, mark_safe
 
+from grandchallenge.pages.models import Page
 from grandchallenge.pages.substitutions import Substitution
+from tests.factories import PageFactory
 
 
 @pytest.mark.parametrize(
@@ -60,3 +62,16 @@ def test_argument_substitution(inp, output):
     )
     assert s.sub(inp) == output
     assert isinstance(s.sub(inp), type(inp))
+
+
+def test_google_group():
+    p = Page(html="{% google_group 'my-group' %}")
+    rendered = p.cleaned_html()
+    assert 'data-groupname="my-group"' in rendered
+
+
+@pytest.mark.django_db
+def test_project_statistics():
+    p = PageFactory(html="{% project_statistics %}")
+    html = p.cleaned_html()
+    assert "data-geochart" in html
